backend/requirements.txt
fastapi
uvicorn
openai
pydantic
httpx
beautifulsoup4
python-multipart
pypdf
python-docx
python-dotenv
